feat(client): server picker on the register screen and pre-flight URL validation
Hoists the existing login-screen server-switch dialog into a shared
ServerSubtitle widget and drops it onto the register screen header so a
self-host user can point at their own server BEFORE creating an account
on the default one (#1063).

Adds an /api/server-info pre-flight to the custom-URL submit path — a
typo no longer fires the logout-on-switch transaction.
