+++ apps/client/lib/src/screens/login_screen.dart
@@ -10,6 +10,7 @@ import '../theme/echo_theme.dart';
 import '../utils/version_utils.dart';
 import '../widgets/auth/auth_layout.dart';
 import '../widgets/auth/auth_scaffold_chrome.dart';
+import '../widgets/auth/server_subtitle.dart';
 import '../widgets/echo_logo_icon.dart';
 import '../widgets/window_chrome.dart';
 
@@ -242,7 +243,7 @@ class _LoginScreenState extends ConsumerState<LoginScreen>
           style: Theme.of(context).textTheme.bodyMedium?.copyWith(fontSize: 16),
         ),
         const SizedBox(height: 6),
-        _ServerSubtitle(serverUrl: serverUrl),
+        ServerSubtitle(serverUrl: serverUrl),
       ],
     );
   }
@@ -323,246 +324,6 @@ class _LoginScreenState extends ConsumerState<LoginScreen>
               : const Text('Log in'),
         ),
       ),
-    );
-  }
-}
-
-/// Small subtitle row shown under the tagline: "Server: echo-messenger.us ›"
-///
-/// Tapping it opens an inline dialog to switch to a known server or enter a
-/// custom URL. This surfaces which backend the user is hitting so a login
-/// failure on the wrong server doesn't look like wrong credentials.
-class _ServerSubtitle extends ConsumerStatefulWidget {
-  final String serverUrl;
-
-  const _ServerSubtitle({required this.serverUrl});
-
-  @override
-  ConsumerState<_ServerSubtitle> createState() => _ServerSubtitleState();
-}
-
-class _ServerSubtitleState extends ConsumerState<_ServerSubtitle> {
-  String _host(String url) {
-    final parsed = Uri.tryParse(url);
-    if (parsed != null && parsed.host.isNotEmpty) return parsed.host;
-    return url;
-  }
-
-  Future<void> _showSwitchDialog() async {
-    final urlController = TextEditingController();
-    final servers = ref.read(knownServersProvider);
-
-    await showDialog<void>(
-      context: context,
-      builder: (ctx) => _ServerSwitchDialog(
-        currentUrl: widget.serverUrl,
-        knownServers: servers,
-        urlController: urlController,
-        onSwitch: (url) async {
-          Navigator.of(ctx).pop();
-          await ref.read(serverUrlProvider.notifier).switchTo(url);
-        },
-      ),
-    );
-    urlController.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final host = _host(widget.serverUrl);
-    return Semantics(
-      button: true,
-      label: 'switch server',
-      child: InkWell(
-        borderRadius: BorderRadius.circular(4),
-        onTap: _showSwitchDialog,
-        child: Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 8),
-          child: Row(
-            mainAxisSize: MainAxisSize.min,
-            children: [
-              Text(
-                'Server: $host',
-                style: TextStyle(color: context.textMuted, fontSize: 12),
-              ),
-              const SizedBox(width: 2),
-              Icon(Icons.chevron_right, size: 14, color: context.textMuted),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-/// Dialog shown when the user taps the server subtitle on the login screen.
-///
-/// Lists [knownServers] as quick-switch tiles and provides a text field to
-/// type a custom URL. Safe to use before login — does not navigate away.
-class _ServerSwitchDialog extends ConsumerStatefulWidget {
-  final String currentUrl;
-  final List<KnownServer> knownServers;
-  final TextEditingController urlController;
-  final Future<void> Function(String url) onSwitch;
-
-  const _ServerSwitchDialog({
-    required this.currentUrl,
-    required this.knownServers,
-    required this.urlController,
-    required this.onSwitch,
-  });
-
-  @override
-  ConsumerState<_ServerSwitchDialog> createState() =>
-      _ServerSwitchDialogState();
-}
-
-class _ServerSwitchDialogState extends ConsumerState<_ServerSwitchDialog> {
-  bool _switching = false;
-
-  String _host(String url) {
-    final parsed = Uri.tryParse(url);
-    if (parsed != null && parsed.host.isNotEmpty) return parsed.host;
-    return url;
-  }
-
-  Future<void> _switch(String url) async {
-    setState(() => _switching = true);
-    try {
-      await widget.onSwitch(url);
-    } finally {
-      if (mounted) setState(() => _switching = false);
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return AlertDialog(
-      backgroundColor: context.surface,
-      shape: RoundedRectangleBorder(
-        borderRadius: BorderRadius.circular(12),
-        side: BorderSide(color: context.border),
-      ),
-      title: Text(
-        'Switch server',
-        style: TextStyle(
-          color: context.textPrimary,
-          fontSize: 18,
-          fontWeight: FontWeight.w600,
-        ),
-      ),
-      content: SizedBox(
-        width: 360,
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          crossAxisAlignment: CrossAxisAlignment.stretch,
-          children: [
-            if (widget.knownServers.isNotEmpty) ...[
-              Text(
-                'Known servers',
-                style: TextStyle(
-                  color: context.textSecondary,
-                  fontSize: 12,
-                  fontWeight: FontWeight.w600,
-                ),
-              ),
-              const SizedBox(height: 8),
-              for (final s in widget.knownServers)
-                _ServerTile(
-                  host: _host(s.url),
-                  isActive: s.url == widget.currentUrl,
-                  onTap: _switching || s.url == widget.currentUrl
-                      ? null
-                      : () => _switch(s.url),
-                ),
-              const SizedBox(height: 16),
-            ],
-            Text(
-              'Custom URL',
-              style: TextStyle(
-                color: context.textSecondary,
-                fontSize: 12,
-                fontWeight: FontWeight.w600,
-              ),
-            ),
-            const SizedBox(height: 8),
-            TextField(
-              controller: widget.urlController,
-              style: TextStyle(color: context.textPrimary, fontSize: 14),
-              decoration: InputDecoration(
-                hintText: 'https://your-server.example.com',
-                hintStyle: TextStyle(color: context.textMuted, fontSize: 13),
-                border: OutlineInputBorder(
-                  borderRadius: BorderRadius.circular(8),
-                ),
-                contentPadding: const EdgeInsets.symmetric(
-                  horizontal: 12,
-                  vertical: 10,
-                ),
-              ),
-              onSubmitted: (v) {
-                final trimmed = v.trim();
-                if (trimmed.isNotEmpty) _switch(trimmed);
-              },
-            ),
-          ],
-        ),
-      ),
-      actions: [
-        TextButton(
-          onPressed: _switching ? null : () => Navigator.of(context).pop(),
-          child: const Text('Cancel'),
-        ),
-        FilledButton(
-          onPressed: _switching
-              ? null
-              : () {
-                  final trimmed = widget.urlController.text.trim();
-                  if (trimmed.isNotEmpty) _switch(trimmed);
-                },
-          child: _switching
-              ? const SizedBox(
-                  width: 16,
-                  height: 16,
-                  child: CircularProgressIndicator(strokeWidth: 2),
-                )
-              : const Text('Connect'),
-        ),
-      ],
-    );
-  }
-}
-
-class _ServerTile extends StatelessWidget {
-  final String host;
-  final bool isActive;
-  final VoidCallback? onTap;
-
-  const _ServerTile({
-    required this.host,
-    required this.isActive,
-    required this.onTap,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return ListTile(
-      dense: true,
-      contentPadding: const EdgeInsets.symmetric(horizontal: 4),
-      leading: Icon(
-        isActive ? Icons.radio_button_checked : Icons.radio_button_off,
-        size: 18,
-        color: isActive ? context.accent : context.textMuted,
-      ),
-      title: Text(
-        host,
-        style: TextStyle(
-          color: context.textPrimary,
-          fontSize: 14,
-          fontWeight: isActive ? FontWeight.w600 : FontWeight.normal,
-        ),
-      ),
-      onTap: onTap,
     );
   }
 }
